search-action

+++ pages/products/[id].js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect, useContext } from "react"
 import Link from "next/link"
 import { useRouter } from "next/router"
+import Swal from "sweetalert2"
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
 import { es } from "date-fns/locale"
 
@@ -171,11 +172,23 @@ export default function Product() {
 
   /* Delete Product */
   const deleteProduct = async () => {
-    try {
-      await firebase.db.collection("products").doc(id).delete()
-      return router.push("/")
-    } catch (error) {
-      console.error("Hubo un error", error)
+    const alert = await Swal.fire({
+      title: "¿Estas seguro?",
+      text: "Un producto que se elimina no se puede recuperar",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+    })
+    if (alert.isConfirmed) {
+      try {
+        await firebase.db.collection("products").doc(id).delete()
+        return router.push("/")
+      } catch (error) {
+        console.error("Hubo un error", error)
+      }
     }
   }
 
@@ -215,19 +228,19 @@ export default function Product() {
 
           {user && productOwner(user.uid) && (
             <div className="action-btn">
-              <Button onClick={deleteProduct} bgColor="#ea0000" hover="#f34336">
-                Eliminar
-              </Button>
-
               <Link
                 href="/products/edit/[editProduct]"
                 as={`/products/edit/${id}`}
                 passHref
               >
-                <Button bgColor="#DA552F" ml="10px">
+                <Button bgColor="#DA552F" mr="10px">
                   Editar
                 </Button>
               </Link>
+
+              <Button onClick={deleteProduct} bgColor="#ea0000" hover="#f34336">
+                Eliminar
+              </Button>
             </div>
           )}
         </MainInfoContainer>
